profile cover & profile image changeing done

+++ server/Config/deleteManager.js
@@ -3,19 +3,36 @@ const path = require("path");
 let file = require("fs");
 
 const deleteFile = async (req, res, next) => {
-	const getFilePath = path.resolve(
-		`../client/public/uploads/profile-img/${req.currentUser.profile_img}`
-	);
+	if (
+		req.currentUser.profile_img === "default-profile.png" ||
+		req.currentUser.cover_img === "default-cover.png"
+	) {
+		next();
+	} else {
+		const whichOne = req.query.whichOne;
 
-	// const getFilePath = `./build/uploads/profile-img/${req.currentUser.profile_img}`;
+		const getFilePath =
+			whichOne === "cover"
+				? path.resolve(
+						`../client/public/uploads/profile-img/${req.currentUser.cover_img}`
+				  )
+				: path.resolve(
+						`../client/public/uploads/profile-img/${req.currentUser.profile_img}`
+				  );
 
-	await file.unlink(getFilePath, (error) => {
-		if (error) {
-			next();
-		} else {
-			next();
-		}
-	});
+		// const getFilePath =
+		// 	whichOne === "cover"
+		// 		? `./build/uploads/profile-img/${req.currentUser.cover_img}`
+		// 		: `./build/uploads/profile-img/${req.currentUser.profile_img}`;
+
+		await file.unlink(getFilePath, (error) => {
+			if (error) {
+				next();
+			} else {
+				next();
+			}
+		});
+	}
 };
 
 module.exports = { deleteFile };
